Add message_sender
- code refector
